requirements.txt
asgiref==3.8.1
Django==5.1.5
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.9
inflection==0.5.1
packaging==24.2
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
